[tests] Remove spurious error log in p2p_segwit.py
Since bfabe480c2ef96111ff820409da0d14f1db4ed2e, when wait_until() fails,
an error message is logged to the test framework log. This means that if
wait_until() is called inside a try-except with the expectation that it
will fail, a spurious error message is logged.

wait_until() shouldn't be called with the expectation of failure. Fix
that in p2p_segwit.py.