[addrman] Fix new table bucketing during unserialization
An addrman entry can appear in up to 8 new table buckets. We store this
entry->bucket indexing during shutdown so that on restart we can restore
the entries to their correct buckets.

Commit f6d2e76c6265d7ca3e684c7819166d9cbdb32a5a broke the
deserialization code so that each entry could only be put in up to one
new bucket. Fix that.