Make CAddrman::Select_ select buckets, not positions, first
The original CAddrMan behaviour (before commit
5888b84b68274e120768551b8f798896d26e1519) was to pick a uniformly
random non-empty bucket, and then pick a random element from that
bucket. That commit, which introduced deterministic placement
of entries in buckets, changed this to picking a uniformly random
non-empty bucket position instead.

This commit reverts the original high-level behavior, but in the
deterministic placement model.