ci: avoid running git diff after patching
Drop `git diff` command so it is easier to run CI locally if git checkout is a
worktree. Currently it fails because the directory is not recognized as a git
repository.

The `git diff` command was added recently in #28359 commit
ccd0064b806761325622d123b5e1a38a433779c0 and can be avoided just by teeing the
patch to stdout